Use a shared debug scope for switch cases
A switch case can have multiple payloads, e.g. in:

```
  case .one(let payload),
       .two(let payload):
    ...
```

We want to emit just one payload debug variable, and to use a shared
debug scope for both patterns in the case.

To achieve that, this patch disallows the creation of multiple
SILDebugScopes which refer to the same location and have the same
context. With this change, two SILDebugScopes are identical iff their
pointers compare equal. The main impact of this is that fewer duplicate
lexical blocks and unnecessary shadow copies are emitted.

This fixes SR-7973, a bug in which "payload" could not be inspected in
lldb, because the DWARF provided multiple locations for the variable in
different scopes.

Also, this reduces the number of SILDebugScope allocations by 20% while
building the swift-stdlib-macosx-x86_64 target. (Apparently it's common
for SILGen to enter the same debug scope multiple times.)

rdar://41048339